requirements.txt
fastapi
uvicorn
streamlit
requests
anthropic
python-dotenv
pydantic
pandas